refactor: split query runtime, field stats, and rebuild helpers
Execute the first substantial cleanup slice from the codebase-wide topology and debt retirement program.\n\nSplit query runtime helpers, embedding stats helpers, field stats collection, action-event rebuild logic, and unified provider-meta harmonization into narrower role-focused modules while keeping the public roots as thin authorities.\n\n🤖 Generated with [Claude Code](https://claude.com/claude-code)\n\nCo-Authored-By: Claude <noreply@anthropic.com>
