+++ tests/unit/core/test_field_stats.py
@@ -18,7 +18,6 @@ from polylogue.schemas.field_stats import (
     is_dynamic_key,
 )
 
-
 # =============================================================================
 # is_dynamic_key
 # =============================================================================
